+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - [Fix] Fix `karafka info` subscription groups count reporting as it was misleading.
 - [Fix] Allow for defining subscription groups with symbols similar to consumer groups and topics to align the API.
 - [Fix] Do not allow for an explicit `nil` as a `subscription_group` block argument.
+- [Fix] Fix instability in subscription groups static members ids when using `--consumer_groups` CLI flag.
+- [Fix] Fix a case in routing, where anonymous subscription group could not be used inside of a consumer group.
 
 ## 20.0.21 (2022-11-25)
 - [Improvement] Make revocation jobs for LRJ topics non-blocking to prevent blocking polling when someone uses non-revocation aware LRJ jobs and revocation happens.